requirements.txt
Cerberus==1.3.4
certifi==2023.5.7
charset-normalizer==3.1.0
Faker==18.7.0
idna==3.4
python-dateutil==2.8.2
PyYAML==6.0
requests==2.30.0
six==1.16.0
urllib3==2.0.2
